utils.tasks (minor): support html in description
