+++ src/appmixer/utils/tasks/RequestApproval/component.json
@@ -45,11 +45,11 @@
                         "tooltip": "The title of the request. This title is displayed in the email that the approver and requester gets as a notification. The title is also displayed in the dashboard that is accessible to the requester and approver."
                     },
                     "description": {
-                        "type": "text",
+                        "type": "textarea",
                         "label": "Description",
                         "group": "transformation",
                         "index": 2,
-                        "tooltip": "The description of the request. This description is displayed in the email that the approver and requester gets as a notification. The description is also displayed in the dashboard that is accessible to the requester and approver."
+                        "tooltip": "The description of the request. This description is displayed in the email that the approver and requester gets as a notification. The description is also displayed in the dashboard that is accessible to the requester and approver.  Markdown is supported. Try <code>*italic*</code>, <code>**bold**</code>, or <code>[link](url)</code>."
                     },
                     "requester": {
                         "type": "text",
